apps/api/requirements.txt
fastapi==0.135.0
uvicorn[standard]==0.41.0
pydantic==2.12.5
pydantic-settings==2.13.1
httpx==0.28.1
redis==7.2.1
supabase==2.28.0
python-dotenv==1.2.1
aiofiles==25.1.0
mistralai==1.12.4
pytest==8.3.4
pytest-asyncio==0.23.6
pytest-mock==3.12.0
